% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDB_enaFuns.R
\name{fetch_enaSeq}
\alias{fetch_enaSeq}
\title{Fetch records and parse them for the requiered information.}
\usage{
fetch_enaSeq(accList, gene, codon_start = F)
}
\description{
`fetch_enaSeq` returns a data.frame with the accession number, the taxid and the region of the sequence for the specified gene (as columns) for the provided accession numbers (as rows).
}
\details{
Uses DBFetch to fetch the records for all provided accession numbers in insdxml format and parse them to return the accession number, the taxid and the region of the sequence for the specified gene.
}
\references{

}
\seealso{
\code{\link{}}

Other ena-db: 
\code{\link{get_enaDetails}()},
\code{\link{get_enaTaxIds}()}
}
\concept{ena-db}
