% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena-db.R
\name{get_enaDetails}
\alias{get_enaDetails}
\title{Get accession numbers and descriptions of records from the ENA databases}
\usage{
get_enaDetails(
  taxIds,
  description = NULL,
  result = "sequence",
  outFile = "./db_downloads/ena_description.json",
  keepFile = T,
  mode = "w",
  overwrite = T
)
}
\arguments{
\item{taxIds}{character string, integer or vector of character string or integer. The taxonomic ids used to include their sub-tree in the search.}

\item{description}{character string or vector of character string or integer. The keywords used to search in the "description" field of the records}

\item{result}{character string. The result type (data set) to search against. Default is "sequence" (Nucleotide sequences). See \url{https://www.ebi.ac.uk/ena/portal/api/results?dataPortal=ena} for all possible results.}

\item{outFile}{character string. The path to where the file should be downloaded. (With this, the URL can't always be read directly, so it's downloaded.)}

\item{keepFile}{logical. Whether to keep or remove the downloaded file once it's been read.}

\item{mode}{character. Whether to write  ("w") over or append ("a") to an existing outFile. Default is "w".}

\item{overwrite}{logical. If the file exist and the mode isn't "a" (append), should the file be overwritten. If set to FALSE, will read the existing file instead of downloading a new one. Default is TRUE.}
}
\description{
`get_enaDetails` returns a data.frame with accession numbers and descriptions (as columns) of all records of a specified database matching specified taxa tree and description.
}
\details{
Uses the ENA Portal API to look up the records from each taxIds taxonomic tree that have the result type and the description keywords specified and return their accession number and description.
}
\references{
\url{https://ena-docs.readthedocs.io/en/latest/retrieval/programmatic-access/taxon-based-search.html}, \url{https://www.ebi.ac.uk/ena/portal/api/}
}
\seealso{
Other ena-db: 
\code{\link{get_enaFasta}()},
\code{\link{get_enaTaxIds}()}
}
\concept{ena-db}
