% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena-db.R
\name{get_enaTaxIds}
\alias{get_enaTaxIds}
\title{Get the taxIds of taxon names from the ENA database}
\usage{
get_enaTaxIds(names, nameType = "scientific")
}
\arguments{
\item{names}{vector of character string. Names of the taxa for which to find their taxId.}

\item{nameType}{character string. The type of name used. Default is 'scientific' Can be set to 'commun' if the commun names should be used.}
}
\description{
`get_taxIds` returns a data.frame with taxId and taxon scientific name as columns.
}
\details{
Uses the ENA API to search the taxonomy database with taxon names and parse the data to return only the taxid and the scientific name for each taxon.
}
\references{

}
\seealso{
\code{\link{https://ena-docs.readthedocs.io/en/latest/retrieval/programmatic-access/taxon-based-search.html}}

Other ena-db: 
\code{\link{get_enaDetails}()},
\code{\link{get_enaFasta}()}
}
\concept{ena-db}
