% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDB_ncbiFuns.R
\name{get_idList}
\alias{get_idList}
\title{================================================================================
Functions to make a database from the NCBI records
================================================================================}
\usage{
get_idList(
  taxa,
  gene = "",
  customTerm = "",
  db = "nucleotide",
  retmode = "json",
  idtype = "acc",
  retmax = "100000"
)
}
\arguments{
\item{taxa}{list of taxa to search for}

\item{gene}{list of the genes/variations of a gene names}

\item{customTerm}{string of terms to add to the eSearch query}

\item{db}{ncbi db to search}

\item{retmode}{return mode of the results; Default is 'json' since this function parse it with jsonlite (could be adapted to parse xml)}

\item{idtype}{type of id; default is accession (UID are depricated for most case)}

\item{retmax}{maximum result returns; Default is 10000 which is the maximum}
}
\value{
a list of accession numbers with the history server information (WebEnv and Query_key)
}
\description{
--------------------------------------------------------------------------------
Get the id list of records matching a query with ESearch
}
\details{
ESearch a NCBI databases for records and returns the accession number.
Uses the ncbi history server to fetch details faster with \code{\link{get_idListDetails}}.
Can also use \code{\link{get_idListDetails}} directly.
NOTE : The ESearch can return up to 100k results. (Will update function to deal with bigger result sets)
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/}, \url{https://www.ncbi.nlm.nih.gov/books/NBK49540/#_chapter4_ESearch_}
arguments check --------------------------------
making the eSearch URL : -----------------------
parsing the data for the idList (accessions) ---
printing number of results ---------------------
}
\seealso{
\code{\link{get_idListDetails}}
}
