% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-sqlite.R
\name{get_namesNodes}
\alias{get_namesNodes}
\title{Download names and nodes files from NCBI}
\usage{
get_namesNodes(outDir = ".")
}
\arguments{
\item{outDir}{the directory to put names.dmp and nodes.dmp in. Required. If directory doesn't exist, it will be created.}
}
\description{
`get_namesNodes` returns a vector of file path strings of the locations of the output files
}
\details{
Download a taxdump.tar.gz file from NCBI servers and extract the names.dmp and nodes.dmp files from it.
These can then be used to create a SQLite database with \code{\link{mk_taxonomy.table}}.
}
\keyword{internal}
