% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-db.R
\name{get_ncbiAcc}
\alias{get_ncbiAcc}
\title{Get the id list of records matching a query with ESearch}
\usage{
get_ncbiAcc(
  taxa,
  gene = "",
  customTerm = "",
  db = "nucleotide",
  retmode = "json",
  retmax = "100000",
  WebEnv = ""
)
}
\arguments{
\item{taxa}{list of taxa to search for}

\item{gene}{list of the genes/variations of a gene names}

\item{customTerm}{string of terms to add to the eSearch query}

\item{db}{ncbi db to search}

\item{retmode}{return mode of the results; Default is 'json' since this function parse it with jsonlite (could be adapted to parse xml)}

\item{retmax}{maximum result returns; Default is 10000 which is the maximum}
}
\description{
`get_ncbiAcc` returns a list of the accession numbers and the history server information (WebEnv and Query_key)
}
\details{
ESearch a NCBI databases for records and returns the accession number.
Can get the details information with \code{\link{get_ncbiDetails}} with the returned list or use \code{\link{get_ncbiDetails}} directly.
}
\note{
The ESearch can return up to 100k results. (Will update function to deal with bigger result sets)
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/}, \url{https://www.ncbi.nlm.nih.gov/books/NBK49540/#_chapter4_ESearch_}
arguments check ------------------------------------------------------------------------------------------------
making the eSearch URL : ---------------------------------------------------------------------------------------
parsing the data for the accList (accessions) -------------------------------------------------------------------
printing number of results -------------------------------------------------------------------------------------
}
\seealso{
\code{\link{get_ncbiDetails}}

Other ncbi-db: 
\code{\link{fetch_ncbiSeq}()},
\code{\link{get_ncbiDetails}()},
\code{\link{get_ncbiFasta}()}
}
\concept{ncbi-db}
