% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-db.R
\name{get_ncbiDetails}
\alias{get_ncbiDetails}
\title{Get the details (description and taxId) from a list of ids with ESummary}
\usage{
get_ncbiDetails(
  taxa,
  gene = "",
  customTerm = "",
  db = "nucleotide",
  retmode = "json",
  retmax = "100000",
  query = NULL
)
}
\arguments{
\item{taxa}{list of taxa to search for; Optional if using the arguments WebEnv, Query_key and nIDs.}

\item{gene}{list of the genes/variations of a gene names; Optional if using the arguments WebEnv, Query_key and nIDs.}

\item{customTerm}{string of terms to add to the eSearch query; Optional if using the arguments WebEnv, Query_key and nIDs.}

\item{db}{ncbi db to search. Default is 'nucleotide'.}

\item{retmode}{return mode of the results; Default is 'json' since this function parse it with jsonlite (could be adapted to parse xml)}

\item{retmax}{maximum result returns; Default is 10000 which is the maximum}

\item{query}{A list of the history server info and the ids associated with it($accession, $Query_key, $WebEnv); Object returned from \code{\link{get_ncbiAcc}}}
}
\description{
`get_ncbiDetails` returns a list containing a list of the history server information ($query) and a list with the fetch details ($details).
}
\details{
Uses ESummary and the history server to get the summary information of NCBI records and parse the data for the desired information (description and taxId).
If \code{\link{get_ncbiAcc}} was used previously, the returned list should be provided as the query argument.
If not, taxa, gene and customTerm should be use. The function will do the ESearch with those first and then get the details.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_ESummary_}
arguments check --------------------------------
getting the records and history server info
getting the records and adding them to the same history server -----------------------------------------------
make sure query is valid
making the eSummary URL ------------------------------------------------------------------------------------------
Function to parse the data ---------------------------------------------------------------------------------------
split if more than 5000 -------------------------------------------------------------------------------------------
server can return up to 10000, but doing so I was missing some records.
Could be tested with more than 5000.
}
\seealso{
\code{\link{get_ncbiAcc}}

Other ncbi-db: 
\code{\link{get_ncbiAcc}()},
\code{\link{get_ncbiFasta}()}
}
\concept{ncbi-db}
