% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi-db.R
\name{get_ncbiFasta}
\alias{get_ncbiFasta}
\title{Get the fasta files from a list of ids with EFetch -------------------------------------------------------------------}
\usage{
get_ncbiFasta(
  accList,
  outDir = ".",
  db = "nucleotide",
  rettype = "fasta_cds_na",
  WebEnv = NULL,
  Query_key = NULL
)
}
\arguments{
\item{accList}{list of taxa to search for; Optional if using the arguments WebEnv, Query_key and nIDs.}

\item{outDir}{list of the genes/variations of a gene names; Optional if using the arguments WebEnv, Query_key and nIDs.}

\item{db}{ncbi db to search. Default is 'nucleotide'.}

\item{rettype}{return type of the results; Default is 'fasta_cds_na' which gets only the CDS. Could be set to fasta to get whole sequences. Note: to get only the gene region, the whole gb file should be fetched and parsed similarly to how it's done for the ENA db;}

\item{WebEnv}{string indicating the Web Environment that contains the id list to be provided as input to EFetch (to be implemented).}

\item{Query_key}{integer specifying which of the id lists attached to the given Web Environment will be used as input to EFetch (to be implemented).}
}
\description{
`get_ncbiFasta` returns a vector of the paths to the downloaded fasta files.
}
\details{
Download EFetched fasta files of the records matching the accList argument and returns a vector of their paths.
History server information could also be use, but has not been implemented yet.
}
\note{
This can only get full sequences or CDS region.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/books/NBK25499/#_chapter4_EFetch_}
makes the ranges of sequences to download (200 at a time) -------------------------------------------------------------
makes file names ------------------------------------------------------------------------------------------------------
make a list of URL to EFetch the fasta files --------------------------------------------------------------------------
downloads the EFetched fasta files ------------------------------------------------------------------------------------
}
\seealso{
\code{\link{get_ncbiDetails}}

Other ncbi-db: 
\code{\link{fetch_ncbiSeq}()},
\code{\link{get_ncbiAcc}()},
\code{\link{get_ncbiDetails}()}
}
\concept{ncbi-db}
