% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDB_ncbiSQLiteFuns.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Parse and merge names.dmp and nodes.dmp into one file (tax.dmp)}
\usage{
get_taxonomy(
  ids,
  sqlFile,
  ranks = c("superkingdom", "phylum", "class", "order", "family", "genus", "species")
)
}
\arguments{
\item{ids}{vector of character string or integer. The taxIds for which to find the taxonomic information.}

\item{sqlFile}{character string. The path to the sqlite database containing the taxonomy table.}

\item{ranks}{vector of character string. The desired ranks to get the taxonomic information of.}
}
\description{
`get_taxonomy` returns a data frame with the names of the desired ranks (columns) corresponding to each taxids (rows).
}
\details{
Uses the taxonomy table in a sqlite file to gets the the parentId of all parents of each taxIds and returns the names corresponding to the ranks desired a data frame with the taxids as row names and ranks as columns.
}
