% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mk_printDF}
\alias{mk_printDF}
\title{Makes the matrix with long row names (like RNA sequences) printable.}
\usage{
mk_printDF(mat, na2char = TRUE, naString = "unidentified")
}
\arguments{
\item{mat}{matrix to be converted. Usually the one returned from dada2::assignedTaxonomy()}

\item{na2char}{Weither to convert NAs to a character string or not; Default TRUE}

\item{naString}{String to use to replace the NA if na2char is set to TRUE; Default is "unidentified"}
}
\description{
`mk_printDF` returns a data.frame without named rows. Optionally, also replaces NAs for a value.
}
\details{
Convert matrix to data.frame and remove row names. Can replace the NA by "NA" or custom value make comparison between tables easier.
}
