% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDB_ncbiSQLiteFuns.R
\name{mk_tax.dmp}
\alias{mk_tax.dmp}
\title{Parse and merge names.dmp and nodes.dmp into one file (tax.dmp)}
\usage{
mk_tax.dmp(namesDmp, nodesDmp)
}
\arguments{
\item{namesDmp}{character string. The path to the names.dmp file.}

\item{nodesDmp}{character string. The path to the nodes.dmp file.}
}
\description{
`mk_tax.dmp` returns a character string of the path to the merged file (tax.dmp).
}
\details{
Helper function for \code{\link{mk_taxonomy.table}}.
Merge names.dmp and nodes.dmp files from NCBI into one tax.dmp file keeping only the columns required (id, name, rank and parentId)
For names.dmp, also only keeps the rows containing the scientific names so there's only one name per id.
}
\keyword{internal}
