% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDB_ncbiSQLiteFuns.R
\name{mk_taxonomy.table}
\alias{mk_taxonomy.table}
\title{Make a taxonomy SQL tables from NCBI nodes.dmp and names.dmp file}
\usage{
mk_taxonomy.table(
  dmpDir = ".",
  dbFile = "./ncbi-db",
  overwrite = FALSE,
  rmTaxDmp = TRUE
)
}
\arguments{
\item{dmpDir}{character string. The path to the directory where to find or download the names.dmp and nodes.dmp files. If directory doesn't exist, it will be created. If the directory is empty, it will ask if you want to download the files.}

\item{dbFile}{character string. The path where the output SQLite file should be saved. Default is './ncbi-db/'. If directory doesn't exist, it will be created.}

\item{overwrite}{logical. If TRUE, overwrites the taxonomy table. Default is FALSE.}

\item{rmTaxDmp}{logical. If FALSE, the merged file created (tax.dmp) won't be removed after the table is created. Default is TRUE.}
}
\description{
`mk_taxonomyTable` returns a character string of the path to the SQLite file in which the table was created.
}
\details{
If the files needed aren't already downloaded, it will download a taxdump.tar.gz file from NCBI servers and extract the names.dmp and nodes.dmp files from it.
The files will then be merged into one temporary .dmp file keeping only the required data (id, scientific name, rank and parentId).
This file is then used to create a SQLite table for easy taxonomy assignment with \code{\link{get_taxonomy}}.
}
\references{
\url{ftp://ftp.ncbi.nih.gov/pub/taxonomy/}, \url{https://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/}
}
\seealso{
\code{\link{get_namesNodes}}, \code{\link{get_taxonomy}}
}
