% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing-records.R
\name{parse_INSDxml}
\alias{parse_INSDxml}
\title{Parse records of INSDxml format for the required information.}
\usage{
parse_INSDxml(
  URL,
  gene,
  gene_type,
  codon_start = F,
  full_seq = F,
  save2csv = F,
  outCsv = NULL
)
}
\arguments{
\item{gene}{character string or vector of character string. The gene for which to find the sequence region.}

\item{codon_start}{logical. Must be set to T if the gene has reading frames so it can adjust the region start accordingly. Default is false.}

\item{full_seq}{logical. Whether to fetch the full sequence or only the gene region. Default is FALSE, gets the gene region.}

\item{save2csv}{logical. Whether to save returned data.frame of the parsed records to a csv file or not. Done using \code{\link{write_parsed2csv}}.}

\item{accList}{character string, integer or vector of character string or integer. The accession numbers used to fetch the records.}

\item{outFile}{character string. The path to the csv file to be written. Optional. If save2csv is set to TRUE and outFile is missing, the file will be saved to './db_downloads/parsedRecords/parsedRecords.Sys.Date.###.csv'}
}
\description{
`parse_INSDxml` parses a INSDxml file and returns a data.frame with the details of the records (as columns) for the provided accession numbers (as rows).
}
\details{
Reads a INSDxml file from a file path or a URL and parse it for the accession number, taxid, description, gene, product, protein id, translation exception, note and sequence fields.
Is used by \code{\link{get_enaFasta}}
}
\seealso{
\code{\link{parse_EMBLxml}}, \code{\link{parse_flatFile}}, \code{\link{get_enaFasta}}
}
