% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savingFiles.R
\name{save_records}
\alias{save_records}
\title{Download the fetched records instead of reading them directly}
\usage{
save_records(URL, outFile = NULL, ext = c("xml", "txt"))
}
\arguments{
\item{URL}{url. Url of the data to save.}

\item{outFile}{character string. The path to where the file should be saved. Optional. If missing, the file will be saved to './db_downloads/records/records_[Sys.Date()]_[###].csv'}

\item{ext}{character string. The extension of the file to save. Chosen depending on format by the fetch function calling it.}
}
\description{
`save_records` downloads a file from a URL and returns a character string of its path.
}
\details{
Downloads a file from a URL and returns a character string of its path.
If no path provided for outFile, file will be saved to "./db_downloads/records/records.Sys.Date.###.csv"
Is used by  \code{\link{parse_EMBLxml}}, \code{\link{parse_flatFile}}, \code{\link{parse_INSDxml}} when save2csv is set to TRUE.
}
