% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{saveRecords}
\alias{saveRecords}
\title{Download the fetched records instead of reading them directly}
\usage{
saveRecords(URL, outFile, ext = c("xml", "txt"))
}
\arguments{
\item{outFile}{character string. The path to the csv file to be written. Optional. If missing, the file will be saved to './db_downloads/records/records_[Sys.Date()]_[###].csv'}

\item{parsedInfo}{data.frame. The data.frame returned from \code{\link{parse_EMBLxml}}, \code{\link{parse_flatFile}} or \code{\link{parse_INSDxml}}.}
}
\description{
`saveRecords` downloads a file from a URL and returns a character string of its path.
}
\details{
Downloads a file from a URL and returns a character string of its path.
If no path provided for outFile, file will be saved to "./db_downloads/records/records.[Sys.Date()].csv"
Is used by  \code{\link{parse_EMBLxml}}, \code{\link{parse_flatFile}}, \code{\link{parse_INSDxml}} when save2csv is set to TRUE.
}
