% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_fasta}
\alias{write_fasta}
\title{=================================================================================
Misc functions
=================================================================================}
\usage{
write_fasta(
  sequences,
  ids = NULL,
  outDir = ".",
  outFile = paste0("sequences", Sys.Date(), ".fasta"),
  nbChar = NULL,
  open = "w"
)
}
\arguments{
\item{sequences}{vector or named list of character string. The sequences to be writen in the file.}

\item{ids}{vector of character string. The ids for the sequences; Optional if a named list is provided}

\item{outDir}{character string. The directory in which to write the file}

\item{outFile}{character string. The name of the file to write}

\item{nbChar}{integer. The number of sequence characters per line. Default is NULL, sequence will be one line}

\item{open}{character string. A description of how to open the connection. Default is "w" for write, could be set to a to append the sequences.}
}
\value{
returns a string with the path to the fasta file.
}
\description{
--------------------------------------------------------------------------------
Write a fasta file from sequences and their names
--------------------------------------------------------------------------------
}
