% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ena-db.R
\name{fetch_enaSeq}
\alias{fetch_enaSeq}
\title{Fetch records and parse them for the required information.}
\usage{
fetch_enaSeq(
  accList,
  gene = gene,
  codon_start = codon_start,
  full_seq = F,
  db = "ena_sequence",
  format = c("embl", "emblxml-1.1", "insdxml"),
  saveRec = F,
  outRec = NULL,
  saveParsedRec = F,
  outParsedRec = NULL
)
}
\arguments{
\item{accList}{character string, integer or vector of character string or integer. The accession numbers used to fetch the records.}

\item{gene}{character string or vector of character string. The gene for which to find the sequence region.}

\item{codon_start}{logical. Must be set to T if the gene has reading frames so it can adjust the region start accordingly. Default is false.}

\item{full_seq}{logical. Whether to fetch the full sequence or only the gene region. Default is FALSE, gets the gene region.}

\item{format}{character string. The format of the fetched file. Supported choice for parsing : 'embl' (flat file), 'emblxml-1.1' (EMBL XML) or 'insdxml' (INSDSeq XML).}
}
\description{
`fetch_enaSeq` returns a data.frame with the accession number, the taxid and the region of the sequence for the specified gene (as columns) for the provided accession numbers (as rows).
}
\details{
Uses DBFetch to fetch the records for all provided accession numbers in specified format and parse them to return the accession number, the taxid and the region of the sequence for the specified gene(s).
}
\references{
\url{http://www.ebi.ac.uk/Tools/dbfetch/dbfetch/dbfetch.databases#ena_sequence}, \url{http://www.ebi.ac.uk/Tools/dbfetch/dbfetch/dbfetch}
}
\seealso{
Other ena-db: 
\code{\link{get_enaDetails}()},
\code{\link{get_enaTaxIds}()}
}
\concept{ena-db}
