% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makingDB_miscFuns.R
\name{mk_printDF}
\alias{mk_printDF}
\title{--------------------------------------------------------------------------------
Makes the matrix returned from dada2::assignedTaxonomy() printable
--------------------------------------------------------------------------------}
\usage{
mk_printDF(mat, na2char = TRUE, naString = "unidentified")
}
\arguments{
\item{mat}{matrix to be converted. Usually the one returned from dada2::assignedTaxonomy()}

\item{na2char}{Weither to convert NAs to a character string or not; Default TRUE}

\item{naString}{String to use to replace the NA if na2char is set to TRUE; Default is "unidentified"}
}
\value{
a data.frame without named rows and, optionnaly, without NAs
}
\description{
Convert matrix to dataframe and remove rownames.
Optionnaly, replace the NA by "NA" to be able to compare between tables
}
